backend/requirements-image-embed.txt
# For scripts/embed_openclip_images.py (OpenCLIP ViT-L/14 image embeddings)
open-clip-torch
torch
torchvision
pillow
requests
